refactor: move all errors and protocols to a single file
