+++ backend/src/presentation/controllers/signup.ts
@@ -1,10 +1,6 @@
-import { InvalidParamError } from '../errors/invalid-param-error'
-import { MissingParamError } from '../errors/missing-param-error'
+import { InvalidParamError, MissingParamError } from '../errors'
 import { badRequest, serverError } from '../helpers/http-helper'
-import { Controller } from '../protocols/controller'
-import { EmailValidator } from '../protocols/email-validator'
-import { FieldValidator } from '../protocols/field-validator'
-import { HttpRequest, HttpResponse } from '../protocols/http'
+import { Controller, EmailValidator, FieldValidator, HttpRequest, HttpResponse } from '../protocols'
 
 export class SignUpController implements Controller {
   constructor (
