+++ backend/src/presentation/protocols/index.ts
@@ -1,0 +1,4 @@
+export * from './controller'
+export * from './email-validator'
+export * from './field-validator'
+export * from './http'
